+++ app/templates/components/admin_header.html
@@ -12,7 +12,10 @@
             </div>
         </a>
         <nav class="header__nav">
-            <a href="/databases" class="header__nav-item">Databases</a>
+            <a href="/admin/teams" class="header__nav-item">Team Management</a>
+        
+    
+            <a href="/admin/org/databases" class="header__nav-item">Databases</a>
         </nav>
         <div class="header__user-actions">
             <!-- <div id="header__notification-button" class="header__notification-container">
