--- app/templates/pages/success.html
@@ -1,9 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}성공 - DataSquare{% endblock %}
-
-{% block content %}
-<div class="success-container">
-    <h1 class="success-message">성공했습니다!</h1>
-</div>
-{% endblock %}
